ipfs_api/requirements.txt
CairoSVG
fastapi
aiohttp
python-dotenv
uvicorn
